Initial Upload
Right now on first run the program crashes with a seg-fault, I could probably fix this in like 2 minutes but I'm really lazy and just wanted to get this uploaded for motivational purposes. It'll only do it the once though, I think it's because of the pokedex.bin generation. Oh well.