Pokedex.c v1.10.10
Fixed pkdxDisplay so that less if statements are required to properly format the text.
Edited code to be more verbose, namely in function names and descriptions. Renamed 'display' to 'list'.
Force pkdxGen to create .bin files, for nice and easy saveData management.
Fixed error where not inputting a maximum range for pkdxUpdate forces a segmentation fault.
Fixed bug where pkdxUpdate can be called for values outside of 1-807 (allowing arbitrary RW to saveData).
Adapted code to handle global variable saveDataFilename, to prevent messing with argv[].
Repurposed pkdxRead to take an array and a filename, to make sure memory is set properly.